Moved models from main app into quiz folder models.py
